BE review fixes 3
